bake-decouple: move serve dev-server option building into bake
